+++ translations/bt_de.ts
@@ -8381,7 +8381,7 @@ Das endgültige Volumen in der Hauptgärung beträgt %1.</translation>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <source>(See Boil tab below)</source>
+        <source>See Boil tab below</source>
         <translation type="unfinished"></translation>
     </message>
 </context>