+++ CHANGES.markdown
@@ -23,9 +23,11 @@ Bug fixes and minor enhancements.
 * Crash on Windows when opening Print and Print Preview dialog [873](https://github.com/Brewtarget/brewtarget/issues/873)
 * Crash when you click Yes or No in the pop-up about downloading the latest version [878](https://github.com/Brewtarget/brewtarget/issues/878)
 * Drop down menu for Style and Equipment are so narrow, cannot see contents [879](https://github.com/Brewtarget/brewtarget/issues/879)
+* OG doesn't immediately change when adding malt to a recipe [880](https://github.com/Brewtarget/brewtarget/issues/880)
+* IBU doesn't update when adding hops [881](https://github.com/Brewtarget/brewtarget/issues/881)
 
 ### Release Timestamp
-Sun, 17 Nov 2024 04:00:11 +0100
+Tue, 19 Nov 2024 04:00:11 +0100
 
 ## v4.0.10
 Bug fixes and minor enhancements.