+++ CHANGES.markdown
@@ -25,7 +25,7 @@ Bug fixes and minor enhancements.
 * Drop down menu for Style and Equipment are so narrow, cannot see contents [879](https://github.com/Brewtarget/brewtarget/issues/879)
 * OG doesn't immediately change when adding malt to a recipe [880](https://github.com/Brewtarget/brewtarget/issues/880)
 * IBU doesn't update when adding hops [881](https://github.com/Brewtarget/brewtarget/issues/881)
-* Errors when loading translations [891](https://github.com/Brewtarget/brewtarget/pull/891)
+* Errors when loading translations [889](https://github.com/Brewtarget/brewtarget/pull/889)
 * Debian package depencies [890](https://github.com/Brewtarget/brewtarget/issues/890)
 
 ### Release Timestamp